refactor: SquareDaoImpl.find() 내부구조 변경
