--- src/main/java/chess/dao/SquareDaoInterface.java
@@ -1,16 +1,0 @@
-package chess.dao;
-
-import chess.model.piece.Piece;
-import chess.model.position.Position;
-import java.util.Map;
-
-public interface SquareDaoInterface {
-
-    void save(Position position, Piece piece);
-
-    Map<String, String> find();
-
-    void delete();
-
-    void update(String position, Piece piece);
-}
