+++ src/main/java/chess/dao/TurnDaoImpl.java
@@ -1,0 +1,61 @@
+package chess.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class TurnDaoImpl implements TurnDao {
+
+    private final Connection connection;
+
+    public TurnDaoImpl() {
+        this.connection = ConnectionManager.getConnection();
+    }
+
+    public void save(final String team) {
+        final String sql = "insert into turn (team) values (?)";
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql);
+            statement.setString(1, team);
+            statement.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public String find() {
+        final String sql = "select team from turn";
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql);
+            final ResultSet resultSet = statement.executeQuery();
+            resultSet.next();
+            return resultSet.getString("team");
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        throw new IllegalArgumentException("[ERROR] 일치하는 데이터가 없습니다.");
+    }
+
+    public void delete() {
+        final String sql = "delete from turn";
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql);
+            statement.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void update(String nowTurn, String nextTurn) {
+        final String sql = "update turn set team = ? where team = ?";
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql);
+            statement.setString(1, nextTurn);
+            statement.setString(2, nowTurn);
+            statement.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+}
