--- src/main/java/chess/dao/TurnDaoInterface.java
@@ -1,12 +1,0 @@
-package chess.dao;
-
-public interface TurnDaoInterface {
-
-    void save(String team);
-
-    String find();
-
-    void delete();
-
-    void update(String nowTurn, String nextTurn);
-}
