+++ src/test/java/chess/dao/FakeTurnDao.java
@@ -1,6 +1,6 @@
 package chess.dao;
 
-public class FakeTurnDao implements TurnDaoInterface {
+public class FakeTurnDao implements TurnDao {
 
     private String team;
